requirements.txt
Django==1.9.6
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
markdown2==2.3.1
Pillow==3.1.0
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter